feat: correct news.ts
